MI5-107 Premješten endpoint u AuthenticationController i logika u AuthenticationService.
