+++ src/main/java/foi/air/szokpt/accountmng/services/AuthenticationService.java
@@ -1,7 +1,9 @@
 package foi.air.szokpt.accountmng.services;
 
+import foi.air.szokpt.accountmng.dtos.respones.TokenValidationResponse;
 import foi.air.szokpt.accountmng.entitites.User;
 import foi.air.szokpt.accountmng.exceptions.AuthenticationException;
+import foi.air.szokpt.accountmng.exceptions.JwtException;
 import foi.air.szokpt.accountmng.repositories.UserRepository;
 import foi.air.szokpt.accountmng.util.JwtUtil;
 import foi.air.szokpt.accountmng.util.hashing.Hasher;
@@ -33,5 +35,16 @@ public class AuthenticationService {
         }
 
         return jwtUtil.generateToken(username, registeredUser.getRole().getName());
+    }
+
+    public TokenValidationResponse validateToken(String authorizationHeader) {
+        String token = jwtUtil.extractToken(authorizationHeader);
+        boolean isValid = jwtUtil.verifyToken(token);
+
+        if (!isValid)
+            throw new JwtException("Invalid token");
+
+        String role = jwtUtil.getRoleName(token);
+        return new TokenValidationResponse(role);
     }
 }
